% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c.index0}
\alias{c.index0}
\title{Combine zero-indexed vectors}
\usage{
\method{c}{index0}(...)
}
\arguments{
\item{...}{objects to be concatenated. All \code{NULL} entries are dropped.}
}
\value{
A zero-indexed vector of class \link{index0}.
}
\description{
When combining vectors, if the first argument to \code{c()} is zero-indexed, then
the result will be zero-indexed as well.
Otherwise, the output will revert to default \R behaviour of indexing from 1.
}
\examples{
x <- as.index0(1:5)
y <- as.index0(6:10)
c(x, y)
c(1:5, y)

}
\seealso{
\code{\link[base:c]{base::c()}}
}
