% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.index0}
\alias{str.index0}
\title{Print Zero-Indexed Values}
\usage{
\method{print}{index0}(x, ...)

\method{str}{index0}(object, ...)
}
\arguments{
\item{x, object}{An object to inspect/print, of class \code{index0}.}

\item{...}{Other arguments, passed to generic function}
}
\description{
When printing zero-indexed objects, it only seems fair that the printed
output is zero-indexed as well. So we replace those little numbers in square
brackets so that they start from zero.
}
\note{
Not yet implemented for matrices, arrays or data frames.
}
\seealso{
\link[base:print]{base::print}, \link[utils:str]{utils::str}
}
